Fixed some templated stuff

+++ Retia/Mathematics/DoubleMathProvider.cs
@@ -10,7 +10,7 @@ using Retia.RandomGenerator;
 
 namespace Retia.Mathematics
 {
-    public partial class DoubleMathProvider : MathProviderBase<double>
+    public partial class DoubleMathProvider : MathProviderBase<Double>
     {
 		[DllImport("FastFuncs")]
         private static extern void ApplySigmoid2D(IntPtr a, IntPtr b, int n);
@@ -22,9 +22,9 @@ namespace Retia.Mathematics
         private static extern void CalculateHD(IntPtr H, IntPtr hCandidate, IntPtr z, IntPtr lastH, int n);
 
         [DllImport("FastFuncs")]
-        private static extern void GravesRMSPropUpdateD(double weightDecay, double learningRate, double decayRate, double momentum, IntPtr weightMatrix, IntPtr grad1_cache, IntPtr grad2_cache, IntPtr momentum_cache, IntPtr gradient, int n);
+        private static extern void GravesRMSPropUpdateD(Double weightDecay, Double learningRate, Double decayRate, Double momentum, IntPtr weightMatrix, IntPtr grad1_cache, IntPtr grad2_cache, IntPtr momentum_cache, IntPtr gradient, int n);
 
-        public override List<int> SoftMaxChoice(Matrix<double> p, double T = 1)
+        public override List<int> SoftMaxChoice(Matrix<Double> p, double T = 1)
         {
             var probs = new List<int>(p.ColumnCount);
             var rnd = SafeRandom.Generator;
@@ -49,48 +49,48 @@ namespace Retia.Mathematics
             return probs;
         }
 
-        public override double Scalar(float scalar)
+        public override Double Scalar(float scalar)
         {
-            return (double)scalar;
+            return (Double)scalar;
         }
 
-        public override double Scalar(double scalar)
+        public override Double Scalar(double scalar)
         {
-            return (double)scalar;
+            return (Double)scalar;
         }
 
-        public override double NaN()
+        public override Double NaN()
         {
-            return double.NaN;
+            return Double.NaN;
         }
 
-        public override void GravesRmsPropUpdate(float weightDecay, float learningRate, float decayRate, float momentum, NeuroWeight<double> weight)
+        public override void GravesRmsPropUpdate(float weightDecay, float learningRate, float decayRate, float momentum, NeuroWeight<Double> weight)
         {
-            using (var ptrs = new MatrixPointers<double>(weight.Weight, weight.Cache1, weight.Cache2, weight.CacheM, weight.Gradient))
+            using (var ptrs = new MatrixPointers<Double>(weight.Weight, weight.Cache1, weight.Cache2, weight.CacheM, weight.Gradient))
             {
                 GravesRMSPropUpdateD(weightDecay, learningRate, decayRate, momentum, ptrs[0], ptrs[1], ptrs[2], ptrs[3], ptrs[4], weight.Weight.Length());
             }
         }
 
-        public override void CalculateH(Matrix<double> H, Matrix<double> hCandidate, Matrix<double> z, Matrix<double> lastH)
+        public override void CalculateH(Matrix<Double> H, Matrix<Double> hCandidate, Matrix<Double> z, Matrix<Double> lastH)
         {
-            using (var ptrs = new MatrixPointers<double>(H, hCandidate, z, lastH))
+            using (var ptrs = new MatrixPointers<Double>(H, hCandidate, z, lastH))
             {
                 CalculateHD(ptrs[0], ptrs[1], ptrs[2], ptrs[3], H.Length());
             }
         }
 
-        public override Matrix<double> SoftMaxNorm(Matrix<double> y, double T = 1)
+        public override Matrix<Double> SoftMaxNorm(Matrix<Double> y, double T = 1)
         {
             var p = y.CloneMatrix();
 
             var ya = y.AsColumnMajorArray();
             var pa = p.AsColumnMajorArray();
 
-            var sums = new double[y.ColumnCount];
+            var sums = new Double[y.ColumnCount];
             for (int i = 0; i < ya.Length; i++)
             {
-                pa[i] = (double)Math.Exp(pa[i] / T);
+                pa[i] = (Double)Math.Exp(pa[i] / T);
                 var c = i / y.RowCount;
                 sums[c] += pa[i];
             }
@@ -104,23 +104,23 @@ namespace Retia.Mathematics
             return p;
         }
 
-        public override void ApplySigmoid2(Matrix<double> matrix1, Matrix<double> matrix2)
+        public override void ApplySigmoid2(Matrix<Double> matrix1, Matrix<Double> matrix2)
         {
-            using (var ptrs = new MatrixPointers<double>(matrix1, matrix2))
+            using (var ptrs = new MatrixPointers<Double>(matrix1, matrix2))
             {
                 ApplySigmoid2D(ptrs[0], ptrs[1], matrix1.Length());
             }
         }
 
-        public override void ApplyTanh(Matrix<double> matrix)
+        public override void ApplyTanh(Matrix<Double> matrix)
         {
-            using (var ptrs = new MatrixPointers<double>(matrix))
+            using (var ptrs = new MatrixPointers<Double>(matrix))
             {
                 ApplyTanhD(ptrs[0], matrix.Length());
             }
         }
 
-        public override double CrossEntropy(Matrix<double> p, Matrix<double> target)
+        public override double CrossEntropy(Matrix<Double> p, Matrix<Double> target)
         {
             if (p.ColumnCount != target.ColumnCount || p.RowCount != target.RowCount)
                 throw new Exception("Matrix dimensions must agree!");
@@ -133,7 +133,7 @@ namespace Retia.Mathematics
             double err = 0.0d;
             for (int i = 0; i < pa.Length; i++)
             {
-                if (double.IsNaN(ta[i]))
+                if (Double.IsNaN(ta[i]))
                     continue;
                 err += ta[i] * Math.Log(pa[i]);
             }
@@ -141,7 +141,7 @@ namespace Retia.Mathematics
             return -err / p.ColumnCount;
         }
 
-        public override double MeanSquare(Matrix<double> y, Matrix<double> target)
+        public override double MeanSquare(Matrix<Double> y, Matrix<Double> target)
         {
             if (y.ColumnCount != target.ColumnCount || y.RowCount != target.RowCount)
                 throw new Exception("Matrix dimensions must agree!");
@@ -155,7 +155,7 @@ namespace Retia.Mathematics
             notNan = 0;
             for (int i = 0; i < ya.Length; i++)
             {
-                if (double.IsNaN(ta[i]))
+                if (Double.IsNaN(ta[i]))
                     continue;
                 notNan++;
                 double delta = ta[i] - ya[i];
@@ -165,22 +165,22 @@ namespace Retia.Mathematics
             return notNan == 0 ? 0.0 : 0.5 * err / notNan;
         }
 
-        protected override Matrix<double> PropagateSingleError(Matrix<double> y, Matrix<double> target, int batchSize)
+        protected override Matrix<Double> PropagateSingleError(Matrix<Double> y, Matrix<Double> target, int batchSize)
         {
-            return target.Map2((targetVal, yVal) => double.IsNaN(targetVal) ? 0.0d : yVal - targetVal, y).Divide(batchSize);
+            return target.Map2((targetVal, yVal) => Double.IsNaN(targetVal) ? (Double)0.0f : yVal - targetVal, y).Divide(batchSize);
         }
 
-        protected override bool AlmostEqual(double a, double b)
+        protected override bool AlmostEqual(Double a, Double b)
         {
             return Math.Abs(a - b) < 10e-10;
         }
 
-        public override double[] Array(params float[] input)
+        public override Double[] Array(params float[] input)
         {
-            return input.Select(x => (double)x).ToArray();
+            return input.Select(x => (Double)x).ToArray();
         }
 
-        public override void ClampMatrix(Matrix<double> matrix, double min, double max)
+        public override void ClampMatrix(Matrix<Double> matrix, Double min, Double max)
         {
             var arr = matrix.AsColumnMajorArray();
             for (int i = 0; i < arr.Length; i++)
@@ -192,13 +192,13 @@ namespace Retia.Mathematics
             }
         }
 
-        public override Matrix<double> RandomMatrix(int rows, int cols, float min, float max)
+        public override Matrix<Double> RandomMatrix(int rows, int cols, float min, float max)
         {
             var random = SafeRandom.Generator;
-            var arr = new double[rows * cols];
+            var arr = new Double[rows * cols];
             for (int i = 0; i < arr.Length; i++)
-                arr[i] = (double)random.NextDouble(min, max);
-            return Matrix<double>.Build.Dense(rows, cols, arr);
+                arr[i] = (Double)random.NextDouble(min, max);
+            return Matrix<Double>.Build.Dense(rows, cols, arr);
         }
 	}
 }
